added conext created database: -AddMigration -update-Database handled Error: FOREIGN KEY constraint may cause or multiple cascade paths

+++ signalRTest_chat/DataLayer/Entities/Roles/RolePermission.cs
@@ -10,10 +10,10 @@ namespace DataLayer.Entities.Roles
 {
     public class RolePermission : BaseEntity
     {
-        public int RoleId { get; set; }
+        public long RoleId { get; set; }
         public Permission Permission { get; set; }
 
-        #region Relation
+        #region Relations
         [ForeignKey("RoleId")]
         public Role Role { get; set; }
         #endregion
